Update to upstream version '1.4'
with Debian dir 0b0d636e76769b309abb838da9361d95c611ebfe

+++ client/internal.h
@@ -2,10 +2,11 @@
  * Internal support functions for the wallet client.
  *
  * Written by Russ Allbery <eagle@eyrie.org>
- * Copyright 2007, 2008, 2010
+ * Copyright 2018 Russ Allbery <eagle@eyrie.org>
+ * Copyright 2007-2008, 2010
  *     The Board of Trustees of the Leland Stanford Junior University
  *
- * See LICENSE for licensing terms.
+ * SPDX-License-Identifier: MIT
  */
 
 #ifndef CLIENT_INTERNAL_H
@@ -41,7 +42,7 @@ struct options {
     char *server;
     char *principal;
     char *user;
-    int port;
+    unsigned short port;
 };
 
 BEGIN_DECLS
